+++ kleiner-brauhelfer/widgets/wdghopfengabe.h
@@ -31,7 +31,7 @@ private slots:
     void on_tbAnteilProzent_valueChanged(double value);
     void on_btnMengeKorrektur_clicked();
     void on_btnAnteilKorrektur_clicked();
-    void on_tbMenge_valueChanged(int value);
+    void on_tbMenge_valueChanged(double value);
     void on_tbKochdauer_valueChanged(int dauer);
     void on_cbZeitpunkt_currentIndexChanged(int index);
     void on_btnLoeschen_clicked();
